Fix connect API to Proxy
Move related helpers for apiUmbrella template
Removed unrelated helpers from common template
Get proxyHost to apiUmbrella template
Change _.get function to _.has function because _.has function returns boolean value instead of value
Set correct 'path' parameter in _.has function
