+++ proxy_backends/client/form/form.html
@@ -28,7 +28,7 @@ https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence
     <!-- API Umbrella -->
     {{# if equals proxy.type "apiUmbrella" }}
 
-      {{> apiUmbrellaProxyForm proxy=proxy apiUmbrella=apiUmbrella api=api }}
+      {{> apiUmbrellaProxyForm proxy=proxy proxyHost=proxyHost api=api }}
 
     {{/ if }}
